Add files via upload
I developed a password generator using java. The program prompts the user to enter the length, whether it should include uppercase, lowercase letters, numbers and special characters.
The user input is then passed to the 'generate password' method which constructs the password based on the provided criteria. The generated password is stored in a StringBuilder and then returned as a string. In case of an empty character set, an exception is thrown.
Overall, the code provides a customizable way to generate random passwords based on the user's preferences.